add sql dir.
