+++ core/src/main/java/io/seata/core/store/db/sql/lock/H2LockStoreSql.java
@@ -13,29 +13,28 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.core.store.db.lock;
+package io.seata.core.store.db.sql.lock;
 
 import io.seata.common.loader.LoadLevel;
 
 /**
- * the database lock store postgre sql
+ * the database lock store H2 sql
  *
  * @author zhangchenghui.dev@gmail.com
  * @since 1.2.0
  */
-@LoadLevel(name = "postgresql")
-public class PostgresqlLockStoreSql extends AbstractLockStoreSql {
+@LoadLevel(name = "h2")
+public class H2LockStoreSql extends AbstractLockStoreSql {
 
     /**
-     * The constant INSERT_LOCK_SQL_POSTGRESQL.
+     * The constant INSERT_LOCK_SQL_H2.
      */
-    private static final String INSERT_LOCK_SQL_POSTGRESQL = "insert into " + LOCK_TABLE_PLACE_HOLD + "(" + ALL_COLUMNS + ")"
-        +
+    private static final String INSERT_LOCK_SQL_H2 = "insert into " + LOCK_TABLE_PLACE_HOLD + "(" + ALL_COLUMNS + ")" +
         "values (?, ?, ?, ?, ?, ?, ?, now(), now())";
 
     @Override
     public String getInsertLockSQL(String lockTable) {
-        return INSERT_LOCK_SQL_POSTGRESQL.replace(LOCK_TABLE_PLACE_HOLD, lockTable);
+        return INSERT_LOCK_SQL_H2.replace(LOCK_TABLE_PLACE_HOLD, lockTable);
     }
 
 }
