+++ core/src/main/java/io/seata/core/store/db/sql/lock/LockStoreSql.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.core.store.db.lock;
+package io.seata.core.store.db.sql.lock;
 
 /**
  * the database lock store sql interface
