+++ core/src/main/java/io/seata/core/store/db/sql/lock/OceanbaseLockStoreSql.java
@@ -13,28 +13,27 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.core.store.db.lock;
+package io.seata.core.store.db.sql.lock;
 
 import io.seata.common.loader.LoadLevel;
 
 /**
- * the database lock store mysql sql
+ * the database lock store oceanbase sql
  *
  * @author zhangchenghui.dev@gmail.com
  * @since 1.2.0
  */
-@LoadLevel(name = "mysql")
-public class MysqlLockStoreSql extends AbstractLockStoreSql {
+@LoadLevel(name = "oceanbase")
+public class OceanbaseLockStoreSql extends AbstractLockStoreSql {
 
     /**
-     * The constant INSERT_LOCK_SQL_MYSQL.
+     * The constant INSERT_LOCK_SQL_OCEANBASE.
      */
-    private static final String INSERT_LOCK_SQL_MYSQL = "insert into " + LOCK_TABLE_PLACE_HOLD + "(" + ALL_COLUMNS + ")" +
+    private static final String INSERT_LOCK_SQL_OCEANBASE = "insert into " + LOCK_TABLE_PLACE_HOLD + "(" + ALL_COLUMNS + ")" +
         "values (?, ?, ?, ?, ?, ?, ?, now(), now())";
 
     @Override
     public String getInsertLockSQL(String lockTable) {
-        return INSERT_LOCK_SQL_MYSQL.replace(LOCK_TABLE_PLACE_HOLD, lockTable);
+        return INSERT_LOCK_SQL_OCEANBASE.replace(LOCK_TABLE_PLACE_HOLD, lockTable);
     }
-
 }
