+++ core/src/main/java/io/seata/core/store/db/sql/lock/OracleLockStoreSql.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.core.store.db.lock;
+package io.seata.core.store.db.sql.lock;
 
 import io.seata.common.loader.LoadLevel;
