+++ core/src/main/resources/META-INF/services/io.seata.core.store.db.sql.LockStoreSqls
@@ -1,5 +1,5 @@
-io.seata.core.store.db.lock.MysqlLockStoreSql
-io.seata.core.store.db.lock.OracleLockStoreSql
-io.seata.core.store.db.lock.OceanbaseLockStoreSql
-io.seata.core.store.db.lock.PostgresqlLockStoreSql
-io.seata.core.store.db.lock.H2LockStoreSql
+io.seata.core.store.db.sql.lock.MysqlLockStoreSql
+io.seata.core.store.db.sql.lock.OracleLockStoreSql
+io.seata.core.store.db.sql.lock.OceanbaseLockStoreSql
+io.seata.core.store.db.sql.lock.PostgresqlLockStoreSql
+io.seata.core.store.db.sql.lock.H2LockStoreSql
